Refactor datastore purge commands
