+++ modules/dkan_datastore/dkan_datastore.services.yml
@@ -71,6 +71,7 @@ services:
       - '@dkan.metastore.storage'
       - '@dkan.datastore.service'
       - '@dkan.datastore.logger_channel'
+  Drupal\dkan_datastore\Service\ResourcePurger: '@dkan.datastore.service.resource_purger'
 
   dkan.datastore.service.factory.import:
     class: \Drupal\dkan_datastore\Service\Factory\ImportServiceFactory
