+++ modules/dkan_datastore/drush.services.yml
@@ -12,23 +12,11 @@ services:
       - '@state'
     tags:
       - { name: drush.command }
-  dkan_datastore.purger.commands:
-    class: Drupal\dkan_datastore\Commands\PurgeCommands
-    arguments:
-      - '@dkan.datastore.service.resource_purger'
-    tags:
-      - { name: drush.command }
   dkan_datastore.reimport.commands:
     class: \Drupal\dkan_datastore\Commands\ReimportCommands
     arguments:
       - '@dkan.datastore.service'
       - '@dkan.common.dataset_info'
       - '@dkan.datastore.logger_channel'
-    tags:
-      - { name: drush.command }
-  dkan_datastore.degraded_mode.commands:
-    class: \Drupal\dkan_datastore\Commands\DegradedModeCommands
-    arguments:
-      - '@state'
     tags:
       - { name: drush.command }
